+++ src/embedded_trace_esp32/serial_tracer.cpp
@@ -1,14 +1,21 @@
 #include "serial_tracer.h"
+#include "esp_idf_tid_fn.h"
 #include <cstdio>
 
 namespace et {
 
-static ThreadId default_tid() { return 1; }
+static ThreadId default_tid_fn() {
+#ifdef ESP_PLATFORM
+    return esp_idf_tid_fn();
+#else
+    return 1;
+#endif
+}
 
 SerialTracer::SerialTracer(Print& output, TimestampFn timestamp_fn,
                            ProcessId pid, ThreadIdFn tid_fn)
     : output_(output), timestamp_fn_(timestamp_fn),
-      pid_(pid), tid_fn_(tid_fn ? tid_fn : default_tid) {}
+      pid_(pid), tid_fn_(tid_fn ? tid_fn : default_tid_fn) {}
 
 void SerialTracer::emit_begin(const char* name) {
     TimestampUs ts = timestamp_fn_();